REFACTOR app and complete standalone version
removed the configparser way of reading in the external twitter account
details (keys, secrets etc). was having trouble running the app in
docker with trying to find its 'settings' file and read it in. too much
hassle and ebooks did a better (cleaner?) way of bringing things in with
an import and a lot less fiddly variable assignments.

might need to clean up and remove some debugging lines.